rename handle_cache.rs -> env_registry.rs
