+++ crates/liboxen/src/lmdb.rs
@@ -13,20 +13,20 @@
 //!     `LmdbEnvConfig` with a fixed per-store sparse map size (no runtime resize).
 //!   * [`lmdb_db`] — [`LmdbDb`], a named sub-database with raw-bytes keys and opaque-bytes values
 //!     copied out as `bytes::Bytes`, plus `open_db` to open one database in its own write txn.
-//!   * [`handle_cache`] — [`LmdbEnvRegistry`] (the ONE path-keyed, weak-retention env cache that
+//!   * [`env_registry`] — [`LmdbEnvRegistry`] (the ONE path-keyed, weak-retention env cache that
 //!     enforces "at most one live env per canonical path") plus [`open_shared_env`], the
 //!     process-global entry point a store calls instead of standing up its own registry.
 //!   * [`store`] — the `LmdbStore` default-method lifecycle trait: an implementor supplies the
 //!     env + database hook and inherits `read`/`write`/`snapshot_to`; its domain ops stay inherent.
 
-pub mod handle_cache;
+pub mod env_registry;
 pub mod lmdb_db;
 pub mod lmdb_env;
 pub mod lmdb_error;
 pub mod store;
 pub mod txn;
 
-pub use handle_cache::{LmdbEnvRegistry, open_shared_env, shared_env_is_live};
+pub use env_registry::{LmdbEnvRegistry, open_shared_env, shared_env_is_live};
 pub use lmdb_db::{LmdbDb, open_db};
 pub use lmdb_env::{LmdbEnv, LmdbEnvConfig};
 pub use lmdb_error::LmdbLayerError;
